==> Code/01_Data_collection.ipynb <==
import pandas as pd
import numpy as np
import requests
import regex as re
from bs4 import BeautifulSoup
from tweetscrape.search_tweets import TweetScrapperSearch
from tweetscrape.profile_tweets import TweetScrapperProfile

setting topic and number of tweets to scrape

topic = 'coronavirus'
number = 40

Scraping the tweets and saving them as a csv to twitter.csv in Data folder

tweet_scrapper = TweetScrapperSearch(topic, num_tweets = number, tweet_dump_path='../Data/twitter.csv', tweet_dump_format = 'csv')
tweet_count, tweet_id, tweet_time, dump_path = tweet_scrapper.get_search_tweets()
print("Extracted {0} tweets till {1} at {2}".format(tweet_count, tweet_time, dump_path))

# EXTRACTING DATA FROM 4 BIG INTERNATIONALISTS

# 1. BBC 

### BBC. Getting all the data from the BBC search

bbc_url = 'https://www.bbc.co.uk/search?q=coronavirus'
bbc = []
res = requests.get(bbc_url)
soup = BeautifulSoup(res.content, 'html.parser')
bbc_headlines = []

for h in soup.find_all('h1'):
    a = h.find('a')
    bbc_dict = {'headline' : h.text,
                'url' : a.attrs['href']}
    bbc_headlines.append(bbc_dict)

bbc = pd.DataFrame(bbc_headlines)

### BBC. Creating a DataFrame from all the text in each story

stories = []
for url in bbc['url'][1:]:
    res = requests.get(url)
    print(res.status_code)
    soup = BeautifulSoup(res.content, 'lxml')
    content = soup.find('div', {'class' : 'story-body__inner'})
    df = pd.DataFrame(content, columns = ['text'])
    df['source'] = url
    stories.append(pd.DataFrame(df))

### BBC. Merging the dataframe

df = pd.concat(stories)

### BBC. Stripping the html, retrieving text

def html_stripper(x):
    try:
        return x.text
    except:
        pass

df['dum'] = df['text'].apply(lambda x : html_stripper(x))

df = df[~df['dum'].isna()]

df = df[~df['dum'].astype(str).str.startswith('\n')]

df = df[['dum', 'source']]

df

### BBC. Merging the text into one block for each story

source = df.groupby('source')

df['source'].unique()

texts = []
for i in df['source'].unique():
    group = source.get_group(i)
    texts.append(pd.DataFrame({'text' : [group['dum'].str.cat(sep = ' ')], 'url':[i]}))
df_text = pd.concat(texts)

df_text.index = [0,1,2,3,4]

df_text

## 2. AP

### AP. Getting all the data from the AP search

apnews_url = 'https://apnews.com/Coronavirus'
ap =[]
res1 = requests.get(apnews_url)
soup = BeautifulSoup(res1.content, 'html.parser')
ap_headlines = []

for a in soup.find_all('a'):
    for h in a.find_all('h1'):
        ap_dict = {'headline' : h.text,
                   'url' : 'https://apnews.com/Coronavirus' + a.attrs['href']}
        ap_headlines.append(ap_dict)
        
ap = pd.DataFrame(ap_headlines)

ap.head()

### AP. Creating a DataFrame from all the text in each story

stories = []
urls = []
for i in ap['url']:
    print(i)
    res = requests.get(i)
    print(res.status_code)
    soup = BeautifulSoup(res.content, 'lxml')
    content = soup.find('div', {'class': 'Article'})
    stories.append(content)
    urls.append(i)

### AP. Stripping the html, retrieving text

ap_scrapes = pd.DataFrame()
ap_scrapes['text'] = stories
ap_scrapes['source'] = urls
ap_scrapes

ap_scrapes['dum'] = ap_scrapes['text'].apply(lambda x : html_stripper(x))

ap_scrapes = ap_scrapes[~ap_scrapes['dum'].isna()]

ap_scrapes = ap_scrapes[~ap_scrapes['dum'].astype(str).str.startswith('\n')]

ap_scrapes = ap_scrapes[['dum', 'source']]
ap_scrapes

### AP. Merging the text into one block for each story

source = ap_scrapes.groupby('source')

ap_scrapes['source'].unique()

texts = []
for i in ap_scrapes['source'].unique():
    group = source.get_group(i)
    texts.append(pd.DataFrame({'text' : [group['dum'].str.cat(sep = ' ')], 'url':[i]}))
ap_text = pd.concat(texts)

ap_text

## 3. Reuters

### Reuters. Getting all the data from the Reuters search

reunews_url = 'https://www.reuters.com/search/news?blob=coronavirus'

rue =[]

res2 = requests.get(reunews_url)

soup = BeautifulSoup(res2.content, 'html.parser')

rue_headlines = []

for h in soup.find_all('h3'):
    try:
        a = h.find('a')
        rue_dict = {'headline' : h.text,
                    'url' : 'https://www.reuters.com' + a.attrs['href']}
        rue_headlines.append(rue_dict)
    except:
        pass
        
rue = pd.DataFrame(rue_headlines)

rue.head()

### Reuters. Creating a DataFrame from all the text in each story

stories = []
urls = []
for i in rue['url']:
    print(i)
    res2 = requests.get(i)
    print(res2.status_code)
    soup = BeautifulSoup(res2.content, 'lxml')
    content = soup.find('div', {'class': 'StandardArticleBody_body'})
    stories.append(content)
    urls.append(i)

rue_scrapes = pd.DataFrame()
rue_scrapes['text'] = stories
rue_scrapes['source'] = urls
rue_scrapes

### Reuters. Stripping the html, retrieving text

rue_scrapes['dum'] = rue_scrapes['text'].apply(lambda x : html_stripper(x))

rue_scrapes = rue_scrapes[~rue_scrapes['dum'].isna()]

rue_scrapes = rue_scrapes[~rue_scrapes['dum'].astype(str).str.startswith('\n')]

rue_scrapes = rue_scrapes[['dum', 'source']]
rue_scrapes

### Rueters. Merging the text into one block for each story¶

source = rue_scrapes.groupby('source')

rue_scrapes['source'].unique()

texts = []
for i in rue_scrapes['source'].unique():
    group = source.get_group(i)
    texts.append(pd.DataFrame({'text' : [group['dum'].str.cat(sep = ' ')], 'url':[i]}))
rue_text = pd.concat(texts)

rue_text

## 4. UPI 

### UPI. Getting all the data from the UPI search

upinews_url = 'https://www.upi.com/search?ss=coronavirus'
upi =[]
res3 = requests.get(upinews_url)
soup = BeautifulSoup(res3.content, 'html.parser')
upi_headlines = []
row_story = soup.find('div', {"class": "row story list"})
for x in row_story.find_all('a'):
    upi_dict ={'headline' : x.attrs['title'],
            'url' : x.attrs['href']}
    upi_headlines.append(upi_dict)
upi = pd.DataFrame(upi_headlines)

upi.head()

### UPI. Creating a DataFrame from all the text in each story

stories = []
urls = []
for i in upi['url']:
    print(i)
    res = requests.get(i)
    print(res.status_code)
    soup = BeautifulSoup(res.content, 'lxml')
    content = soup.find('article', {'itemprop': 'articleBody'})
    stories.append(content)
    urls.append(i)

upi_scrapes = pd.DataFrame()
upi_scrapes['text'] = stories
upi_scrapes['source'] = urls
upi_scrapes

### UPI. Stripping the html, retrieving text

upi_scrapes['dum'] = upi_scrapes['text'].apply(lambda x : html_stripper(x))
upi_scrapes['dum'] = upi_scrapes['dum'].str.replace('\n', '')
upi_scrapes['dum'] = upi_scrapes['dum'].apply(lambda x:re.sub('.*[\(]UPI[\)]\s--\s', '', x))


upi_scrapes

### UPI. Merging the text into one block for each story¶


source = upi_scrapes.groupby('source')

upi_scrapes['source'].unique()

texts = []
for i in upi_scrapes['source'].unique():
    group = source.get_group(i)
    texts.append(pd.DataFrame({'text' : [group['dum'].str.cat(sep = ' ')], 'url':[i]}))
upi_text = pd.concat(texts)

upi_text

final_df = pd.concat([df_text, ap_text, rue_text, upi_text])

final_df.reset_index().drop(columns = 'index', inplace = True)

final_df.to_csv('../Data/news_df.csv', index = False)
